--- photobooth/services/containers.py
@@ -1,142 +1,0 @@
-"""Containers module."""
-
-
-import logging
-
-from dependency_injector import containers, providers
-
-from .aquisitionservice import AquisitionService
-from .baseservice import BaseService
-from .filtetransferservice import FileTransferService
-from .gpioservice import GpioService
-from .informationservice import InformationService
-from .keyboardservice import KeyboardService
-from .mediacollectionservice import MediacollectionService
-from .mediaprocessingservice import MediaprocessingService
-from .printingservice import PrintingService
-from .processingservice import ProcessingService
-from .shareservice import ShareService
-from .sseservice import SseService
-from .systemservice import SystemService
-from .wledservice import WledService
-
-logger = logging.getLogger(__name__)
-
-
-def init_res_obj_service(_obj_: BaseService, sse_service: SseService, *args):
-    """Initialize services as ressources.
-    Ensure to no reraise exceptions, so only the service will fail instead the
-    whole app crash because of exception not catched
-
-    Args:
-        _obj_ (BaseService): Class of type BaseService or derived from that
-        config (AppConfig): Instance of config
-        sse_service (sse_service)
-
-    Yields:
-        _obj_: Initialized resource (inherited from BaseService)
-    """
-    resource = None
-
-    try:
-        resource = _obj_(sse_service, *args)
-        resource.start()
-    except Exception as exc:
-        logger.exception(exc)
-        logger.critical("could not init/start resource")
-    finally:
-        yield resource
-
-    try:
-        if resource:  # if not none
-            resource.stop()
-    except Exception as exc:
-        logger.exception(exc)
-        logger.critical("could not stop resource")
-
-
-class ServicesContainer(containers.DeclarativeContainer):
-    sse_service = providers.Dependency(instance_of=SseService)
-
-    # Services: Core
-
-    aquisition_service = providers.Resource(
-        init_res_obj_service,
-        AquisitionService,
-        sse_service,
-    )
-
-    information_service = providers.Resource(
-        init_res_obj_service,
-        InformationService,
-        sse_service,
-        aquisition_service,
-    )
-
-    wled_service = providers.Resource(
-        init_res_obj_service,
-        WledService,
-        sse_service,
-    )
-
-    mediaprocessing_service = providers.Singleton(
-        MediaprocessingService,
-        sse_service,
-    )
-    mediacollection_service = providers.Singleton(
-        MediacollectionService,
-        sse_service,
-        mediaprocessing_service,
-    )
-
-    processing_service = providers.Singleton(
-        ProcessingService,
-        sse_service,
-        aquisition_service,
-        mediacollection_service,
-        mediaprocessing_service,
-        wled_service,
-    )
-
-    system_service = providers.Factory(
-        SystemService,
-        sse_service,
-    )
-
-    printing_service = providers.Resource(
-        init_res_obj_service,
-        PrintingService,
-        sse_service,
-        mediacollection_service,
-    )
-
-    keyboard_service = providers.Resource(
-        init_res_obj_service,
-        KeyboardService,
-        sse_service,
-        processing_service,
-        printing_service,
-        mediacollection_service,
-    )
-
-    gpio_service = providers.Resource(
-        init_res_obj_service,
-        GpioService,
-        sse_service,
-        processing_service,
-        printing_service,
-        mediacollection_service,
-    )
-
-    share_service = providers.Resource(
-        init_res_obj_service,
-        ShareService,
-        sse_service,
-        mediacollection_service,
-    )
-
-    filetransfer_service = providers.Resource(
-        init_res_obj_service,
-        FileTransferService,
-        sse_service,
-    )
